Rename the Lua library files to be more correct
