+++ Sprocket/CMakeLists.txt
@@ -79,8 +79,8 @@ add_library(Sprocket STATIC
             Utility/Yaml.cpp
 
             Scripting/LuaScript.cpp
-            Scripting/LuaLibrary.cpp
-            Scripting/LuaMaths.cpp
+            Scripting/lua_registry.cpp
+            Scripting/lua_maths.cpp
             Scripting/LuaConverter.cpp
             Scripting/lua_events.cpp)
